requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
httpx==0.27.2
streamlit==1.44.1
pypdf==5.4.0
python-docx==1.1.2